LayoutTests/imported/w3c/web-platform-tests/workers/baseurl/alpha/xhr-in-sharedworker-expected.txt

CONSOLE MESSAGE: line 7: ReferenceError: Can't find variable: SharedWorker

FAIL Base URL in classic shared workers: XHR ReferenceError: Can't find variable: SharedWorker
